requirements.txt
Flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
redis==5.0.7
urllib3==2.2.2